refactor show_name_helpers -> show.naming
