+++ medusa/failed_processor.py
@@ -15,7 +15,8 @@
 #
 # You should have received a copy of the GNU General Public License
 
-from . import app, logger, show_name_helpers
+from medusa.show import naming
+from . import app, logger
 from .helper.exceptions import FailedPostProcessingFailedException
 from .name_parser.parser import InvalidNameException, InvalidShowException, NameParser
 from .search.queue import FailedQueueItem
@@ -42,7 +43,7 @@ class FailedProcessor(object):
         """
         self._log(u'Failed download detected: ({nzb}, {dir})'.format(nzb=self.nzb_name, dir=self.dir_name))
 
-        releaseName = show_name_helpers.determine_release_name(self.dir_name, self.nzb_name)
+        releaseName = naming.determine_release_name(self.dir_name, self.nzb_name)
         if not releaseName:
             self._log(u'Warning: unable to find a valid release name.', logger.WARNING)
             raise FailedPostProcessingFailedException()
