+++ medusa/post_processor.py
@@ -22,18 +22,15 @@ import os
 import re
 import stat
 import subprocess
-
 from collections import OrderedDict
 
 import adba
-
 import rarfile
-
 from rarfile import Error as RarError, NeedFirstVolume
-
 from six import text_type
 
-from . import app, common, db, failed_history, helpers, history, logger, notifiers, show_name_helpers
+from medusa.show import naming
+from . import app, common, db, failed_history, helpers, history, logger, notifiers
 from .helper.common import episode_num, pretty_file_size, remove_extension
 from .helper.exceptions import (EpisodeNotFoundException, EpisodePostProcessingFailedException,
                                 ShowDirectoryNotFoundException)
@@ -1143,7 +1140,7 @@ class PostProcessor(object):
                 sql_l.append(cur_ep.get_sql())
 
         # Just want to keep this consistent for failed handling right now
-        nzb_release_name = show_name_helpers.determine_release_name(self.folder_path, self.nzb_name)
+        nzb_release_name = naming.determine_release_name(self.folder_path, self.nzb_name)
         if nzb_release_name is not None:
             failed_history.log_success(nzb_release_name)
         else:
